// src/enums/enums.common.ts
export enum Frontend {
  CoreTech = "Core Tech",
  Frameworks = "Frameworks",
  StateManagement = "State Management",
  Styling = "Styling",
  UILibraries = "UI Libraries",
  FormManagement = "Form Management",
  Animation = "Animation",
}

export enum BackEnd {
  Frameworks = "Frameworks",
  Databases = "Databases",
  CMS = "CMS",
  APIManagement = "API Management",
}

export enum Design {
  UXUITools = "UX/UI Tools",
}

export enum Methodologies {
  Workflows = "Workflows",
  ProjectManagement = "Project Management",
}

export enum Testing {
  Frameworks = "Frameworks",
  APItools = "API tools",
}

export enum Documentation {
  Workshops = "Workshops",
}

export enum DevOps {
  Containerization = "Containerization",
}

export enum Versionning {
  VersionControl = "Version Control",
  Repositories = "Repositories",
}